+++ src/index.jsx
@@ -1,4 +1,5 @@
 
-export { Link } from './components/Link.js'
+export { Link, navigate } from './components/Link.js'
 export { Router } from './components/Router.js'
 export { Route } from './components/Route.js'
+export { RouterErrorBoundary } from './components/ErrorBoundary.js'